+++ src/spec/KeyInputSpec.ts
@@ -20,13 +20,13 @@ describe('A keyboard', () => {
   it('should fire keydown events', () => {
     let eventFired = false;
 
-    keyboard.on('down', function(ev: ex.Input.KeyEvent) {
+    keyboard.on('down', function (ev: ex.Input.KeyEvent) {
       if (ev.key === ex.Input.Keys.Up) {
         eventFired = true;
       }
     });
 
-    (<any>mockWindow).emit('keydown', { keyCode: ex.Input.Keys.Up });
+    (<any>mockWindow).emit('keydown', { code: ex.Input.Keys.Up });
 
     expect(eventFired).toBeTruthy();
   });
@@ -34,27 +34,27 @@ describe('A keyboard', () => {
   it('should fire keyup events', () => {
     let eventFired = false;
 
-    keyboard.on('up', function(ev: ex.Input.KeyEvent) {
+    keyboard.on('up', function (ev: ex.Input.KeyEvent) {
       if (ev.key === ex.Input.Keys.Up) {
         eventFired = true;
       }
     });
 
-    (<any>mockWindow).emit('keyup', { keyCode: ex.Input.Keys.Up });
+    (<any>mockWindow).emit('keyup', { code: ex.Input.Keys.Up });
 
     expect(eventFired).toBeTruthy();
   });
 
   it('should know if keys are pressed', () => {
     // push key down
-    (<any>mockWindow).emit('keydown', { keyCode: ex.Input.Keys.Up });
+    (<any>mockWindow).emit('keydown', { code: ex.Input.Keys.Up });
 
     expect(keyboard.isHeld(ex.Input.Keys.Up)).toBeTruthy();
     expect(keyboard.wasReleased(ex.Input.Keys.Up)).toBeFalsy();
     expect(keyboard.wasPressed(ex.Input.Keys.Up)).toBeTruthy();
 
     // release key
-    (<any>mockWindow).emit('keyup', { keyCode: ex.Input.Keys.Up });
+    (<any>mockWindow).emit('keyup', { code: ex.Input.Keys.Up });
 
     expect(keyboard.getKeys().length).toBe(0);
     expect(keyboard.isHeld(ex.Input.Keys.Up)).toBeFalsy();
@@ -63,13 +63,13 @@ describe('A keyboard', () => {
 
   it('should have keys stay held until released', () => {
     // push key down
-    (<any>mockWindow).emit('keydown', { keyCode: ex.Input.Keys.Up });
-    (<any>mockWindow).emit('keydown', { keyCode: ex.Input.Keys.Down });
+    (<any>mockWindow).emit('keydown', { code: ex.Input.Keys.Up });
+    (<any>mockWindow).emit('keydown', { code: ex.Input.Keys.Down });
 
     keyboard.update();
 
     // release key
-    (<any>mockWindow).emit('keyup', { keyCode: ex.Input.Keys.Up });
+    (<any>mockWindow).emit('keyup', { code: ex.Input.Keys.Up });
 
     expect(keyboard.wasReleased(ex.Input.Keys.Down)).toBeFalsy();
     expect(keyboard.isHeld(ex.Input.Keys.Up)).toBeFalsy();
@@ -77,7 +77,7 @@ describe('A keyboard', () => {
     keyboard.update();
 
     // release key
-    (<any>mockWindow).emit('keyup', { keyCode: ex.Input.Keys.Down });
+    (<any>mockWindow).emit('keyup', { code: ex.Input.Keys.Down });
 
     expect(keyboard.wasReleased(ex.Input.Keys.Down)).toBeTruthy();
   });