+++ webpack.config.prod.js
@@ -1,4 +1,4 @@
-const merge = require('webpack-merge');
+const { merge } = require('webpack-merge');
 
 const base = require('./webpack.config.js');